tree: merge execution_error.rs into cass_error.rs
The `CassError`-related items were arbitrarily split between two
modules: `cass_error.rs` and `execution_error.rs`.

As this division is artificial, `execution_error.rs` contents have been
moved to `cass_error.rs`.
