+++ scylla-rust-wrapper/src/future.rs
@@ -1,9 +1,6 @@
 use crate::RUNTIME;
 use crate::argconv::*;
-use crate::cass_error::CassError;
-use crate::cass_error::CassErrorMessage;
-use crate::cass_error::ToCassError;
-use crate::execution_error::CassErrorResult;
+use crate::cass_error::{CassError, CassErrorMessage, CassErrorResult, ToCassError as _};
 use crate::prepared::CassPrepared;
 use crate::query_result::{CassNode, CassResult};
 use crate::types::*;
