+++ scylla-rust-wrapper/src/lib.rs
@@ -19,7 +19,6 @@ pub(crate) mod collection;
 pub(crate) mod config_value;
 pub(crate) mod date_time;
 pub(crate) mod exec_profile;
-pub(crate) mod execution_error;
 pub(crate) mod future;
 pub(crate) mod inet;
 #[cfg(cpp_integration_testing)]
